Port some test_shell_tests, and platform cleanup for more of them. Many of these files that I don't added to the build would even compile and link, but don't yet pass with current test_shell on Linux. They just hang.
Patch from Paweł Hajdan jr <phajdan.jr@gmail.com> on
http://codereview.chromium.org/9417

Review URL: http://codereview.chromium.org/9641


git-svn-id: svn://svn.chromium.org/chrome/trunk/src@4892 0039d316-1c4b-4281-b951-d872f2087c98

+++ webkit/glue/mimetype_unittest.cc
@@ -49,7 +49,7 @@ TEST_F(MimeTypeTests, MimeTypeTests) {
     "text/rtf",
     "application/x-javascript",
   };
-  for (int i = 0; i < arraysize(plain_text); ++i) {
+  for (size_t i = 0; i < arraysize(plain_text); ++i) {
     CheckMimeType(plain_text[i], expected_src);
   }
 
@@ -60,7 +60,7 @@ TEST_F(MimeTypeTests, MimeTypeTests) {
     "text/xsl",
     "application/xhtml+xml",
   };
-  for (int i = 0; i < arraysize(html_src); ++i) {
+  for (size_t i = 0; i < arraysize(html_src); ++i) {
     CheckMimeType(html_src[i], L"HTML text");
   }
 
@@ -72,7 +72,7 @@ TEST_F(MimeTypeTests, MimeTypeTests) {
     "image/jpeg",
     "image/bmp",
   };
-  for (int i = 0; i < arraysize(not_text); ++i) {
+  for (size_t i = 0; i < arraysize(not_text); ++i) {
     CheckMimeType(not_text[i], L"");
     test_shell_->webView()->StopLoading();
   }
